+++ src/admin/management-role/ManagementGuru.jsx
@@ -11,7 +11,7 @@ import {
 import { DeleteOutlined } from "@ant-design/icons";
 import CmsTemplate from "../../components/template/CmsTemplate";
 import Loading from "../../components/template/Loading";
-import ModalPopup from "../../components/ConfirmModal";
+import ConfirmModal from "../../components/template/ConfirmModal";
 import Utils from "../../utils/Utils";
 import api from "../../config/axios";
 import { showErrorNotification, showSuccessNotification } from "../../components/template/Notification";
@@ -149,7 +149,7 @@ const ManagementGuru = () => {
                     <Button
                         key={`delete-${record.id}`}
                         onClick={() =>
-                            ModalPopup({
+                            ConfirmModal({
                                 title: "Apakah anda ingin hapus data ini?",
                                 onOk: () => handleDelete(record.id),
                                 content: "Klik Ok untuk hapus data",
